Change package name to get rid of relational since other packages were also refactored
