+++ persistence/google-cloud-spanner/src/main/java/org/apache/polaris/persistence/spanner/DatabaseClientSupplier.java
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-package org.apache.polaris.persistence.relational.spanner;
+package org.apache.polaris.persistence.spanner;
+
+import java.util.function.Supplier;
 
 import com.google.cloud.spanner.DatabaseClient;
-import java.util.function.Supplier;
 
 public interface DatabaseClientSupplier extends Supplier<DatabaseClient> {}
